version 1.0

+++ src/main/java/DomesticatableTrader.java
@@ -33,7 +33,7 @@ class DomesticatableTrader extends Trader<Domesticatable> {
 
     @Override
     public int getSellingPrice(Domesticatable item) {
-        int super_price = super.getSellingPrice(item);
+        int super_price = (int) super.getSellingPrice(item);
 
         if (super_price == Tradable.MISSING_PRICE) {
             return super_price;
